Make sure Makefile.in gets generated properly
config.guess, config.sub, depcomp, etc, are all scripts that ship with
automake, and which will be updated with a new automake version; it's
generally a bad idea to use one of these scripts from a different
automake version.

Moreover, recent versions of automake also contain and require a
"compile" script, which isn't in this directory.

Instead, pass the "--add-missing" parameter to automake, which will copy
the required files to the expected location. This will make automake
work, regardless of the version, and will avoid any issues due to the
used scripts not matching the used automake version.