Implemented tuple extensions to #declare and #local.
A new tuple-style syntax extensions has been added to #declare and #local, in order to simplify passing multiple values out of a macro:
    #declare ( ID1, ID2, ...) = ( EXPR1, EXPR2, ... );
This statement is fully equivalent to:
    #declare ID1 = EXPR1;
    #declare ID1 = EXPR2;
   ...
(Note that this equivalence holds true even for constructs like `#declare (A,B) = (B,A)`, which does _not_ swap the contents of A and B as you might expect.)
The terminating semicolon is optional.

In addition, a similar syntax extension has been added for easier assignment of vector components to individual variables:
    #declare < ID1, ID2, ... > = VECTOR_EXPR;
This statement is fully equivalent to:
    #local TEMP = VECTOR_EXPR;
    #declare ID1 = TEMP.x;
    #declare IDY = TEMP.y;
   ...
except that the new syntax does not actually define a local variable named TEMP.
The terminating semicolon is mandatory.

The #local statement syntax has been extended accordingly.

The `deprecated` keyword can be used with these new syntax variants by placing it right before the respective identifier, e.g.:
    #declare (A, deprecated B) = (47, 11);